requirements.txt
# Web Scraping
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3
selenium==4.15.2
webdriver-manager==4.0.1

# RAG & Embeddings
chromadb==0.4.22
sentence-transformers==2.2.2

# LLM
google-generativeai==0.8.3

# UI
streamlit==1.31.0

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0

